refactor: move constants to separate file, change variable names
