+++ src/components/sort/sortSelect.tsx
@@ -1,21 +1,7 @@
 import { ChangeEvent } from "react";
 import { useSearchParams } from "react-router-dom";
+import { sortOption, SORT_OPTIONS, SORT_PARAM, SORT_TITLE } from "../../constants/sort";
 
-export const SORT_PARAM = 'sort';
-
-export enum SORT_CRITERIA {
-  price = 'price',
-  rating = 'rating',
-  discount = 'discount'
-}
-export enum SORT_ORDER {
-  ASC = 'ASC',
-  DESC = 'DESC'
-}
-
-export type sortOption = `${SORT_CRITERIA}-${SORT_ORDER}`;
-
-export const SORT_OPTIONS: sortOption[] = ['price-ASC', 'price-DESC', 'rating-ASC', 'rating-DESC', 'discount-ASC', 'discount-DESC'];
 
 export default function SortSelect() { 
   const [searchParams, setSearchParams] = useSearchParams();
@@ -27,20 +13,20 @@ export default function SortSelect() {
     })
   }
 
-  const inQuery = searchParams.has(SORT_PARAM) && SORT_OPTIONS.includes(searchParams.get(SORT_PARAM) as sortOption) 
+  const selectedValue = searchParams.has(SORT_PARAM) && SORT_OPTIONS.includes(searchParams.get(SORT_PARAM) as sortOption) 
     ? searchParams.get(SORT_PARAM) as sortOption 
-    : 'sort-title'
+    : SORT_TITLE
 
   return (
     <div className="sort-bar">
       <select 
         onChange={onChangeHandler} 
-        value={inQuery}
+        value={selectedValue}
         name="sort-select" 
         id="sort-select"
         >
-        {['sort-title', ...SORT_OPTIONS].map((el, i) => {
-          if (el === 'sort-title') {
+        {[SORT_TITLE, ...SORT_OPTIONS].map((el, i) => {
+          if (el === SORT_TITLE) {
             return <option key={i} value={el} disabled>{`Sort options:`}</option>;
           }
           return <option key={i} value={el}>{`Sort by ${el.replace('-', ' ')}`}</option>;
